# openapi.yml
openapi: 3.0.0
info:
  version: '0.1'
  title: 'Language'
  description:  This service manages FCBH languages 
  license:
    name: MIT
servers:
  - url: 'localhost:3000'
paths:  
...
  /language/{id}:    
    put:
      summary: Update an existing language
      operationId: updateLanguage
      tags:
        - language
      parameters:
        - name: id
          in: path
          description: language identifier
          required: true
          schema:
              $ref: '#/components/schemas/id'        
      requestBody:
          $ref: '#/components/requestBodies/language-body' 
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: integer
                minimum: 1
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'    
    
    get:
      summary: Returns details about a particular language
      operationId: listLanguage
      tags:
        - language
      parameters:
        - name: id
          in: path
          description: language identifier
          required: true
          schema:
              $ref: '#/components/schemas/id'  
      responses:
        '200':
          description: 'Details about a single language'
          headers:
            x-next:
              description: A link to the next page of responses
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/language'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Deletes an existing language
      operationId: deleteLanguage
      tags:
        - language
      parameters:
        - name: id
          in: path
          description: language identifier
          required: true
          schema:
              $ref: '#/components/schemas/id'  
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: integer
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'                
...
components:
  schemas:
    id:
        description: identifier
        type: integer
        format: int32   
        readOnly: true
        example: 201
    language-code:
        description: FCBH Language Code identifier
        type: string
        minLength: 3
        maxLength: 3
        example: SPA
    iso3:
        description: ISO 693-3 Language Code identifier
        type: string
        minLength: 3
        maxLength: 3  
        example: SPA
    rolv-id:
        description: Registry of Language Variety identifier
        type: string
        minLength: 10
        maxLength: 10  
        example: "04557"
    glotto-id:
        description: GlottoLog identifier
        type: string
        minLength: 10
        maxLength: 10 
        example: "stan1288"
    sensitivity:
        description: enumeration of language sensitivity options
        type: string
        enum: [low, moderate, high, extreme]  
    language-director:
        description: Name of director responsible for language translation
        type: string
        example: "Bob"
...
    language:
      type: object
      required:
        - code
        - name
      properties:
        id:
          $ref: '#/components/schemas/id'  
        code:
          $ref: '#/components/schemas/language-code'
        name:
          type: string
          example: "Spanish"
        iso3:
          $ref: '#/components/schemas/iso3'
        rolv_id:
          $ref: '#/components/schemas/rolv-id'
        glotto_id:
          $ref: '#/components/schemas/glotto-id'
        speakers:
          type: integer
          example: 300000
        sensitivity:
          $ref: '#/components/schemas/sensitivity'          
        language-director:
          $ref: '#/components/schemas/language-director'   
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
  examples:
      language:
        summary: example of a minimal language object
        value:
            code: ENG
            name: English 
...
  requestBodies:
      language-body:
          description: a JSON object containing language information
          required: true
          content: 
              application/json:
                  schema:
                     $ref: '#/components/schemas/language'   
                  example: 
                     $ref: '#/components/examples/language'   
...
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
security:
  - BasicAuth: []